Module arcalog::collection::prow

Structs

Annotations

BuildInfo

Capabilities

Clonerefs

ConfigMap

Container

DecorationConfig

EmptyDir

Env

EnvFrom

ExtraRef

FieldRef

GcsConfiguration

GithubTeamSlug

HostPath

Initupload

Item

Item struct containing each individual Prow job’s metadata

Item2

JobTypeBuilds

Job types with a link to all their build IDs

Labels

Limits

Metadata

NodeSelector

PlaceEntrypoint

PodSpec

PrevReportStates

ProwjobDefaults

Pull

Refs

ReporterConfig

Requests

Requests2

Requests3

Requests4

Requests5

RerunAuthConfig

Resources

Resources2

Root

Root struct from Prow jobs metadata containing a list of all recent jobs

Secret

SecretKeyRef

SecretRef

SecurityContext

SecurityContext2

Sidecar

Slack

Spec

Status

Individual job status metadata

Toleration

UtilityImages

ValueFrom

Volume

VolumeMount

Functions

download_artifacts

Downloads the latest artifacts from Prow for the given build ID

download_artifacts_recursive

Recursively downloads all artifacts for a given url into the given path

download_metadata

Downloads the latest Prow jobs metadata and creates searchable files

find_urls

Find URLs in a text using a regular expression searching for HTML tags

get_build_info

Get build information for a given build ID